backend/requirements.txt
Flask
ollama
supabase
python-dotenv
pandas
flask-cors